Move zen_set into registries/ as set.py
BaseSet/SimpleSet/IndexedSet are the value objects a SetRegistry stores,
so they belong with the runtime containers. Kept as a separate module
(not merged into set_registry.py) because registries/base.py needs
BaseSet and SetRegistry subclasses Registry.

- zen_garden/model/zen_set.py -> zen_garden/model/registries/set.py
- tests/.../test_zen_set.py -> test_set.py

mypy, ruff, black clean; unit tests + spot integration tests pass.
