+++ zen_garden/elements/technology/constraints/technology_diffusion_limit_constraint.py
@@ -7,7 +7,7 @@ import xarray as xr
 from linopy.expressions import LinearExpression
 
 from zen_garden.model.component_types.constraint import GenericConstraint
-from zen_garden.model.zen_set import BaseSet
+from zen_garden.model.registries.set import BaseSet
 
 
 class TechnologyDiffusionLimitConstraint(GenericConstraint):
