+++ zen_garden/model/__init__.py
@@ -4,11 +4,11 @@ from zen_garden.config import Config
 from zen_garden.model.registries.constraint import ConstraintRegistry
 from zen_garden.model.registries.multi_index_helper import MultiIndexHelper
 from zen_garden.model.registries.parameter import DictParameter, ParameterRegistry
+from zen_garden.model.registries.set import BaseSet, IndexedSet, SimpleSet
 from zen_garden.model.registries.set_registry import SetRegistry
 from zen_garden.model.registries.variable import VariableRegistry
 from zen_garden.model.time_steps import TimeStepsDicts
 from zen_garden.model.zen_model import ZenModel
-from zen_garden.model.zen_set import BaseSet, IndexedSet, SimpleSet
 
 __all__ = [
     "Config",
