+++ zen_garden/model/registries/parameter.py
@@ -7,8 +7,8 @@ import pandas as pd
 import xarray as xr
 
 from zen_garden.model.registries.base import Registry
+from zen_garden.model.registries.set import BaseSet
 from zen_garden.model.registries.set_registry import SetRegistry
-from zen_garden.model.zen_set import BaseSet
 
 
 class DictParameter(object):
